implemented several regular user operations

+++ src/main/java/model/Account.java
@@ -4,16 +4,16 @@ import java.sql.Date;
 
 public class Account {
 
-    private int id;
+    private Long id;
     private String type;
     private double balance;
     private Date creationDate;
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
